+++ server/report_listener_udp_helpers_test.go
@@ -23,8 +23,8 @@ func generateTestReport(shortID uint32, timeslot uint32, privKey glow.PrivateKey
 
 // sendUDPReport simulates sending a report to the server via UDP.
 // The server should be listening on the given IP and port.
-func sendUDPReport(report []byte, port int) error {
-	conn, err := net.Dial("udp", fmt.Sprintf("%s:%d", serverIP, port))
+func sendUDPReport(report []byte, port uint16) error {
+	conn, err := net.Dial("udp", fmt.Sprintf("%s:%d", serverIP, int(port)))
 	if err != nil {
 		return err
 	}
